+++ cmd/app/main.go
@@ -1,51 +1,27 @@
 package main
 
 import (
-	"context"
 	"encoding/json"
 	"flag"
 	"fmt"
 	"github.com/jmcgrath207/k8s-ephemeral-storage-metrics/pkg/dev"
 	"github.com/jmcgrath207/k8s-ephemeral-storage-metrics/pkg/node"
+	"github.com/jmcgrath207/k8s-ephemeral-storage-metrics/pkg/pod"
 	"github.com/panjf2000/ants/v2"
 	"github.com/prometheus/client_golang/prometheus"
 	"github.com/prometheus/client_golang/prometheus/promhttp"
-	"github.com/rs/zerolog"
 	"github.com/rs/zerolog/log"
-	v1 "k8s.io/api/core/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/client-go/informers"
-	"k8s.io/client-go/tools/cache"
 	"net/http"
-	"net/http/pprof"
-	"os"
-	"runtime"
 	"strconv"
-	"sync"
 	"time"
 )
 
 var (
-	sampleInterval                                  int64
-	sampleIntervalMill                              int64
-	ephemeralStoragePodUsage                        bool
-	ephemeralStorageContainerLimitsPercentage       bool
-	ephemeralStorageContainerVolumeLimitsPercentage bool
-	podDataWaitGroup                                sync.WaitGroup
-	podGaugeVec                                     *prometheus.GaugeVec
-	containerPercentageLimitsVec                    *prometheus.GaugeVec
-	containerPercentageVolumeLimitsVec              *prometheus.GaugeVec
-	podResourceLookup                               map[string]pod
-	podResourceLookupMutex                          sync.RWMutex
-	Node                                            node.Node
+	sampleInterval     int64
+	sampleIntervalMill int64
+	Node               node.Node
+	Pod                pod.Collector
 )
-
-type volume struct {
-	AvailableBytes int64  `json:"availableBytes"`
-	CapacityBytes  int64  `json:"capacityBytes"`
-	UsedBytes      int    `json:"usedBytes"`
-	Name           string `json:"name"`
-}
 
 type ephemeralStorageMetrics struct {
 	Node struct {
@@ -62,214 +38,8 @@ type ephemeralStorageMetrics struct {
 			UsedBytes      float64 `json:"usedBytes"`
 		} `json:"ephemeral-storage"`
 
-		Volumes []volume `json:"volume,omitempty"`
+		Volumes []pod.Volume `json:"volume,omitempty"`
 	}
-}
-
-type pod struct {
-	containers []container
-}
-
-type container struct {
-	name            string
-	limit           float64
-	emptyDirVolumes []emptyDirVolumes
-}
-
-type emptyDirVolumes struct {
-	name      string
-	mountPath string
-	sizeLimit float64
-}
-
-// Collector for container data
-func getContainerData(c v1.Container, p v1.Pod) container {
-
-	setContainer := container{}
-	setContainer.name = c.Name
-	matchKey := v1.ResourceName("ephemeral-storage")
-
-	if ephemeralStorageContainerVolumeLimitsPercentage && p.Spec.Volumes != nil {
-		collectMounts := false
-
-		podMountsMap := make(map[string]float64)
-		for _, v := range p.Spec.Volumes {
-			if v.VolumeSource.EmptyDir != nil {
-				if v.VolumeSource.EmptyDir.SizeLimit != nil {
-					podMountsMap[v.Name] = v.VolumeSource.EmptyDir.SizeLimit.AsApproximateFloat64()
-					collectMounts = true
-				}
-			}
-
-		}
-
-		if collectMounts {
-			var collectVolumes []emptyDirVolumes
-
-			for _, volumeMount := range c.VolumeMounts {
-				size, ok := podMountsMap[volumeMount.Name]
-				if ok {
-					collectVolumes = append(collectVolumes, emptyDirVolumes{name: volumeMount.Name, mountPath: volumeMount.MountPath, sizeLimit: size})
-				}
-			}
-
-			setContainer.emptyDirVolumes = collectVolumes
-
-		}
-
-	}
-	if ephemeralStorageContainerLimitsPercentage {
-		for key, val := range c.Resources.Limits {
-			if key == matchKey {
-				setContainer.limit = val.AsApproximateFloat64()
-			}
-		}
-	}
-	return setContainer
-}
-
-// Collector for pod data
-func getPodData(p v1.Pod) {
-	if p.Status.Phase == "Running" {
-		var collectContainers []container
-
-		for _, x := range p.Spec.Containers {
-			collectContainers = append(collectContainers, getContainerData(x, p))
-		}
-
-		podResourceLookupMutex.Lock()
-		podResourceLookup[p.Name] = pod{containers: collectContainers}
-		podResourceLookupMutex.Unlock()
-	}
-}
-
-func initGetPodsData() {
-	podResourceLookup = make(map[string]pod)
-	// Init Get List of all pods
-	pods, err := dev.Clientset.CoreV1().Pods("").List(context.TODO(), metav1.ListOptions{})
-	if err != nil {
-		log.Error().Msgf("Error getting pods: %v\n", err)
-		os.Exit(1)
-	}
-
-	for _, p := range pods.Items {
-		getPodData(p)
-	}
-	podDataWaitGroup.Done()
-
-}
-
-func podWatch() {
-	podDataWaitGroup.Wait()
-	stopCh := make(chan struct{})
-	defer close(stopCh)
-	sharedInformerFactory := informers.NewSharedInformerFactory(dev.Clientset, time.Duration(sampleInterval)*time.Second)
-	podInformer := sharedInformerFactory.Core().V1().Pods().Informer()
-
-	// Define event handlers for Pod events
-	eventHandler := cache.ResourceEventHandlerFuncs{
-		UpdateFunc: func(oldObj, newObj interface{}) {
-			p := newObj.(*v1.Pod)
-			getPodData(*p)
-		},
-		DeleteFunc: func(obj interface{}) {
-			p := obj.(*v1.Pod)
-			podResourceLookupMutex.Lock()
-			delete(podResourceLookup, p.Name)
-			podResourceLookupMutex.Unlock()
-			evictPodFromMetrics(*p)
-		},
-	}
-
-	// Register the event handlers with the informer
-	_, err := podInformer.AddEventHandler(eventHandler)
-	if err != nil {
-		log.Err(err)
-		os.Exit(1)
-	}
-
-	// Start the informer to begin watching for Pod events
-	go sharedInformerFactory.Start(stopCh)
-
-	for {
-		time.Sleep(time.Duration(sampleInterval) * time.Second)
-		select {
-		case <-stopCh:
-			log.Error().Msg("Watcher podWatch stopped.")
-			os.Exit(1)
-		}
-	}
-
-}
-
-func evictPodFromMetrics(p v1.Pod) {
-
-	podGaugeVec.DeletePartialMatch(prometheus.Labels{"pod_name": p.Name})
-	for _, c := range p.Spec.Containers {
-		containerPercentageLimitsVec.DeletePartialMatch(prometheus.Labels{"container": c.Name})
-		containerPercentageVolumeLimitsVec.DeletePartialMatch(prometheus.Labels{"container": c.Name})
-	}
-}
-
-func createMetricsValues(podName string, podNamespace string, nodeName string, usedBytes float64, availableBytes float64, capacityBytes float64, volumes []volume) {
-
-	var setValue float64
-	podResourceLookupMutex.RLock()
-	podResult, okPodResult := podResourceLookup[podName]
-	podResourceLookupMutex.RUnlock()
-
-	Node.SetMetrics(nodeName, availableBytes, capacityBytes)
-
-	// TODO: something seems wrong about the metrics.
-	//		the volume capacityBytes is not reflected in this query
-	// 		kubectl get --raw "/api/v1/nodes/ephemeral-metrics-cluster-worker/proxy/stats/summary"
-	// 		need to source it from the pod spec
-	// 		make issue upstream with CA advisor
-	// TODO: need to do a grow and shrink test for this.
-	if ephemeralStorageContainerVolumeLimitsPercentage {
-		// TODO: what a mess...need to figure out a better way.
-		if okPodResult {
-			for _, c := range podResult.containers {
-				if c.emptyDirVolumes != nil {
-					for _, edv := range c.emptyDirVolumes {
-						for _, v := range volumes {
-							if edv.name == v.Name {
-								labels := prometheus.Labels{"pod_namespace": podNamespace,
-									"pod_name": podName, "node_name": nodeName, "container": c.name, "volume_name": v.Name,
-									"mount_path": edv.mountPath}
-								containerPercentageVolumeLimitsVec.With(labels).Set((float64(v.UsedBytes) / edv.sizeLimit) * 100.0)
-							}
-						}
-					}
-				}
-			}
-		}
-	}
-
-	if ephemeralStorageContainerLimitsPercentage {
-		if okPodResult {
-			for _, c := range podResult.containers {
-				labels := prometheus.Labels{"pod_namespace": podNamespace,
-					"pod_name": podName, "node_name": nodeName, "container": c.name}
-				if c.limit != 0 {
-					// Use Limit from Container
-					setValue = (usedBytes / c.limit) * 100.0
-				} else {
-					// Default to Node Available Ephemeral Storage
-					setValue = (availableBytes / capacityBytes) * 100.0
-				}
-				containerPercentageLimitsVec.With(labels).Set(setValue)
-			}
-		}
-	}
-
-	if ephemeralStoragePodUsage {
-		labels := prometheus.Labels{"pod_namespace": podNamespace,
-			"pod_name": podName, "node_name": nodeName}
-		podGaugeVec.With(labels).Set(usedBytes)
-		log.Debug().Msg(fmt.Sprintf("pod %s/%s on %s with usedBytes: %f", podNamespace, podName, nodeName, usedBytes))
-	}
-
 }
 
 func setMetrics(nodeName string) {
@@ -297,8 +67,8 @@ func setMetrics(nodeName string) {
 			log.Warn().Msg(fmt.Sprintf("pod %s/%s on %s has no metrics on its ephemeral storage usage", podName, podNamespace, nodeName))
 			continue
 		}
-		createMetricsValues(podName, podNamespace, nodeName, usedBytes,
-			availableBytes, capacityBytes, p.Volumes)
+		Node.SetMetrics(nodeName, availableBytes, capacityBytes)
+		Pod.SetMetrics(podName, podNamespace, nodeName, usedBytes, availableBytes, capacityBytes, p.Volumes)
 	}
 
 	adjustTime := sampleIntervalMill - time.Now().Sub(start).Milliseconds()
@@ -311,72 +81,10 @@ func setMetrics(nodeName string) {
 
 }
 
-func createMetrics() {
-
-	podGaugeVec = prometheus.NewGaugeVec(prometheus.GaugeOpts{
-		Name: "ephemeral_storage_pod_usage",
-		Help: "Current ephemeral byte usage of pod",
-	},
-		[]string{
-			// name of pod for Ephemeral Storage
-			"pod_name",
-			// namespace of pod for Ephemeral Storage
-			"pod_namespace",
-			// Name of Node where pod is placed.
-			"node_name",
-		},
-	)
-
-	prometheus.MustRegister(podGaugeVec)
-
-	containerPercentageLimitsVec = prometheus.NewGaugeVec(prometheus.GaugeOpts{
-		Name: "ephemeral_storage_container_limit_percentage",
-		Help: "Percentage of ephemeral storage used by a container in a pod",
-	},
-		[]string{
-			// name of pod for Ephemeral Storage
-			"pod_name",
-			// namespace of pod for Ephemeral Storage
-			"pod_namespace",
-			// Name of Node where pod is placed.
-			"node_name",
-			// Name of container
-			"container",
-		},
-	)
-
-	prometheus.MustRegister(containerPercentageLimitsVec)
-
-	containerPercentageVolumeLimitsVec = prometheus.NewGaugeVec(prometheus.GaugeOpts{
-		Name: "ephemeral_storage_container_volume_limit_percentage",
-		Help: "Percentage of ephemeral storage used by a container's volume in a pod",
-	},
-		[]string{
-			// name of pod for Ephemeral Storage
-			"pod_name",
-			// namespace of pod for Ephemeral Storage
-			"pod_namespace",
-			// Name of Node where pod is placed.
-			"node_name",
-			// Name of container
-			"container",
-			// Name of Volume
-			"volume_name",
-			// Name of Mount Path
-			"mount_path",
-		},
-	)
-
-	prometheus.MustRegister(containerPercentageVolumeLimitsVec)
-
-}
-
 func getMetrics() {
 
 	Node.WaitGroup.Wait()
-	if ephemeralStorageContainerLimitsPercentage || ephemeralStorageContainerVolumeLimitsPercentage {
-		podDataWaitGroup.Wait()
-	}
+	Pod.WaitGroup.Wait()
 
 	p, _ := ants.NewPoolWithFunc(Node.MaxNodeQueryConcurrency, func(node interface{}) {
 		setMetrics(node.(string))
@@ -395,65 +103,22 @@ func getMetrics() {
 	}
 }
 
-type LineInfoHook struct{}
-
-func (h LineInfoHook) Run(e *zerolog.Event, l zerolog.Level, msg string) {
-	_, file, line, ok := runtime.Caller(0)
-	if ok {
-		e.Str("line", fmt.Sprintf("%s:%d", file, line))
-	}
-}
-
-func setLogger() {
-	logLevel := dev.GetEnv("LOG_LEVEL", "info")
-	zerolog.TimeFieldFormat = zerolog.TimeFormatUnix
-	level, err := zerolog.ParseLevel(logLevel)
-	if err != nil {
-		panic(err.Error())
-	}
-	zerolog.SetGlobalLevel(level)
-	log.Hook(LineInfoHook{})
-
-}
-
-func enablePprof() {
-	mux := http.NewServeMux()
-	mux.HandleFunc("/debug/pprof/", pprof.Index)
-	mux.HandleFunc("/debug/pprof/cmdline", pprof.Cmdline)
-	mux.HandleFunc("/debug/pprof/profile", pprof.Profile)
-	mux.HandleFunc("/debug/pprof/symbol", pprof.Symbol)
-	mux.HandleFunc("/debug/pprof/trace", pprof.Trace)
-	err := http.ListenAndServe("localhost:6060", mux)
-	if err != nil {
-		log.Error().Msgf("Pprof could not start localhost:")
-	}
-}
-
 func main() {
 	flag.Parse()
 	port := dev.GetEnv("METRICS_PORT", "9100")
-	// TODO: move to a configmap.
-	ephemeralStoragePodUsage, _ = strconv.ParseBool(dev.GetEnv("EPHEMERAL_STORAGE_POD_USAGE", "false"))
-	ephemeralStorageContainerLimitsPercentage, _ = strconv.ParseBool(dev.GetEnv("EPHEMERAL_STORAGE_CONTAINER_LIMIT_PERCENTAGE", "false"))
-	ephemeralStorageContainerVolumeLimitsPercentage, _ = strconv.ParseBool(dev.GetEnv("EPHEMERAL_STORAGE_CONTAINER_VOLUME_LIMITS_PERCENTAGE", "false"))
 
 	// Shared Vars
 	pprofEnabled, _ := strconv.ParseBool(dev.GetEnv("PPROF", "false"))
 	sampleInterval, _ = strconv.ParseInt(dev.GetEnv("SCRAPE_INTERVAL", "15"), 10, 64)
 	sampleIntervalMill = sampleInterval * 1000
 
-	Node = node.NewNode(sampleInterval)
+	dev.SetLogger()
+	dev.SetK8sClient()
+	Node = node.NewCollector(sampleInterval)
+	Pod = pod.NewCollector(sampleInterval)
 
 	if pprofEnabled {
-		go enablePprof()
-	}
-	setLogger()
-	dev.SetK8sClient()
-	createMetrics()
-	if ephemeralStorageContainerLimitsPercentage || ephemeralStorageContainerVolumeLimitsPercentage {
-		podDataWaitGroup.Add(1)
-		go initGetPodsData()
-		go podWatch()
+		go dev.EnablePprof()
 	}
 	go Node.Get()
 	go Node.Watch()
