+++ pkg/dev/util.go
@@ -1,10 +1,15 @@
 package dev
 
 import (
+	"fmt"
+	"github.com/rs/zerolog"
 	"github.com/rs/zerolog/log"
 	"k8s.io/client-go/kubernetes"
 	"k8s.io/client-go/rest"
+	"net/http"
+	"net/http/pprof"
 	"os"
+	"runtime"
 )
 
 var (
@@ -33,4 +38,38 @@ func SetK8sClient() {
 	}
 	log.Debug().Msg("Successful got the in cluster client")
 
+}
+
+type LineInfoHook struct{}
+
+func (h LineInfoHook) Run(e *zerolog.Event, l zerolog.Level, msg string) {
+	_, file, line, ok := runtime.Caller(0)
+	if ok {
+		e.Str("line", fmt.Sprintf("%s:%d", file, line))
+	}
+}
+
+func SetLogger() {
+	logLevel := GetEnv("LOG_LEVEL", "info")
+	zerolog.TimeFieldFormat = zerolog.TimeFormatUnix
+	level, err := zerolog.ParseLevel(logLevel)
+	if err != nil {
+		panic(err.Error())
+	}
+	zerolog.SetGlobalLevel(level)
+	log.Hook(LineInfoHook{})
+
+}
+
+func EnablePprof() {
+	mux := http.NewServeMux()
+	mux.HandleFunc("/debug/pprof/", pprof.Index)
+	mux.HandleFunc("/debug/pprof/cmdline", pprof.Cmdline)
+	mux.HandleFunc("/debug/pprof/profile", pprof.Profile)
+	mux.HandleFunc("/debug/pprof/symbol", pprof.Symbol)
+	mux.HandleFunc("/debug/pprof/trace", pprof.Trace)
+	err := http.ListenAndServe("localhost:6060", mux)
+	if err != nil {
+		log.Error().Msgf("Pprof could not start localhost:")
+	}
 }
